First cut at moving stuff out of EquoListTask and into P2Multitool.
